+++ collector/shadow_parallel.py
@@ -44,15 +44,55 @@ SOURCES = {
             lambda t: P.parse_sporttery(t, "dlt"),
         ),
     ],  # 前5+后2、5位期号→真第二独立源
-    # 以下 6 彩种：17500 官方镜像采集（vs 生产库的逐期比对仍只 ssq/dlt；这里监控“采集健康”）
-    "fc3d": [("17500", "http://data.17500.cn/3d_asc.txt", lambda t: P.parse_17500(t, "fc3d", 3, 0))],
-    "pl3": [("17500", "http://data.17500.cn/pl3_asc.txt", lambda t: P.parse_17500(t, "pl3", 3, 0))],
-    "pl5": [("17500", "http://data.17500.cn/pl5_asc.txt", lambda t: P.parse_17500(t, "pl5", 5, 0))],
-    "qlc": [("17500", "http://data.17500.cn/7lc_asc.txt", lambda t: P.parse_17500(t, "qlc", 7, 1))],
+    # 以下 6 彩种：官方源 + 17500 镜像交叉（vs 生产库的逐期深度比对仍主要 ssq/dlt）
+    "fc3d": [
+        (
+            "cwl",
+            "https://www.cwl.gov.cn/cwl_admin/front/cwlkj/search/kjxx/findDrawNotice?name=3d&issueCount=8",
+            lambda t: P.parse_cwl(t, "fc3d"),
+        ),
+        ("17500", "http://data.17500.cn/3d_asc.txt", lambda t: P.parse_17500(t, "fc3d", 3, 0)),
+    ],
+    "pl3": [
+        (
+            "sporttery",
+            "https://webapi.sporttery.cn/gateway/lottery/getHistoryPageListV1.qry?gameNo=35&provinceId=0&pageSize=8&isVerify=1&pageNo=1",
+            lambda t: P.parse_sporttery(t, "pl3"),
+        ),
+        ("17500", "http://data.17500.cn/pl3_asc.txt", lambda t: P.parse_17500(t, "pl3", 3, 0)),
+    ],
+    "pl5": [
+        (
+            "sporttery",
+            "https://webapi.sporttery.cn/gateway/lottery/getHistoryPageListV1.qry?gameNo=350133&provinceId=0&pageSize=8&isVerify=1&pageNo=1",
+            lambda t: P.parse_sporttery(t, "pl5"),
+        ),
+        ("17500", "http://data.17500.cn/pl5_asc.txt", lambda t: P.parse_17500(t, "pl5", 5, 0)),
+    ],
+    "qlc": [
+        (
+            "cwl",
+            "https://www.cwl.gov.cn/cwl_admin/front/cwlkj/search/kjxx/findDrawNotice?name=qlc&issueCount=8",
+            lambda t: P.parse_cwl(t, "qlc"),
+        ),
+        ("17500", "http://data.17500.cn/7lc_asc.txt", lambda t: P.parse_17500(t, "qlc", 7, 1)),
+    ],
     "qxc": [
-        ("17500", "http://data.17500.cn/7xc_asc.txt", lambda t: P.parse_17500(t, "qxc", 7, 0))
+        (
+            "sporttery",
+            "https://webapi.sporttery.cn/gateway/lottery/getHistoryPageListV1.qry?gameNo=04&provinceId=0&pageSize=8&isVerify=1&pageNo=1",
+            lambda t: P.parse_sporttery(t, "qxc"),
+        ),
+        ("17500", "http://data.17500.cn/7xc_asc.txt", lambda t: P.parse_17500(t, "qxc", 7, 0)),
     ],  # 5 位期号
-    "kl8": [("17500", "http://data.17500.cn/kl8_asc.txt", lambda t: P.parse_17500(t, "kl8", 20, 0))],
+    "kl8": [
+        (
+            "cwl",
+            "https://www.cwl.gov.cn/cwl_admin/front/cwlkj/search/kjxx/findDrawNotice?name=kl8&issueCount=8",
+            lambda t: P.parse_cwl(t, "kl8"),
+        ),
+        ("17500", "http://data.17500.cn/kl8_asc.txt", lambda t: P.parse_17500(t, "kl8", 20, 0)),
+    ],
 }
 
 
@@ -60,7 +100,7 @@ REFERERS = {"sporttery": "https://static.sporttery.cn/"}  # 体彩官方 webapi 
 
 
 def get(url, referer=None):
-    cmd = ["curl", "-k", "--ssl-no-revoke", "--max-time", "60", "-A", UA, "-s"]
+    cmd = ["curl", "--ssl-no-revoke", "--max-time", "60", "-A", UA, "-s"]
     if referer:
         cmd += ["-e", referer]
     cmd.append(url)